+++ packages/rulesets/src/oas/index.ts
@@ -522,6 +522,7 @@ const ruleset = {
       recommended: true,
       formats: [oas2],
       severity: 0,
+      resolved: false,
       given: '$',
       then: {
         function: oasDocumentSchema,
@@ -678,6 +679,7 @@ const ruleset = {
       severity: 0,
       formats: [oas3],
       recommended: true,
+      resolved: false,
       given: '$',
       then: {
         function: oasDocumentSchema,
@@ -705,6 +707,33 @@ const ruleset = {
           checkSubstitutions: true,
           requireDefault: true,
         },
+      },
+    },
+    'oas3-callbacks-in-callbacks': {
+      message: 'Callbacks should not be defined within a callback',
+      formats: [oas3],
+      recommended: true,
+      given: ['#OperationObject.callbacks[*][*][*].callbacks'],
+      then: {
+        function: undefined,
+      },
+    },
+    'oas3_1-servers-in-webhook': {
+      message: 'Servers should not be defined in a webhook.',
+      formats: [oas3_1],
+      recommended: true,
+      given: ['$.webhooks.servers', '$.webhooks[*][*].servers'],
+      then: {
+        function: undefined,
+      },
+    },
+    'oas3_1-callbacks-in-webhook': {
+      message: 'Callbacks should not be defined in a webhook.',
+      formats: [oas3_1],
+      recommended: true,
+      given: ['$.webhooks[*][*].callbacks'],
+      then: {
+        function: undefined,
       },
     },
   },